Major changes to how the code is organized
Almost everything is now a function or part of a function.

The functions to generate rows and squares, for example,
are part of the function 'createGrid'.

Also, changing color is now a function called 'paint'.

The button that asks how many squares per side and stores
value, despite not being a function, it calls more than
one function.

This approach allowed the code to be cleaner and easier
to understand.