Fix button to generate random colors
Instead of create a new unique function
to generate random colors, we use the default
function that already exists to paint.

The difference now is that we have a keyword that
changes it's value if the right button is pressed.

So, if we press the button 'random colors', this
keyword is updated and so is the color.