Fix the button 'shading'
Each mouse pass adds another 10% of black
to it. After 10 passes the square is completely
black.

The core behind this feature was the CSS function
brightness().

We also use JavaScript slice() method to handle
some strings related to the mentioned CSS function.

The JavaScript Number() method was useful to convert
these strings to numbers.

Then, making use of these numbers, we created a function to
deal with the shading called 'makeItDarker'.